// src/types/core/api.ts
export type AppMinManifest = {
  packaging_format: 1 | 2
  id: string
  name: string
  description: Translation
  version: string
  maintainers: string[]
  integration: {
    architectures: string | string[]
    ldap: 'not_relevant' | boolean
    sso: 'not_relevant' | boolean
    multi_instance: boolean
    disk: string
    ram: { build: string; runtime: string }
    yunohost: string
  }
  antifeatures?: Obj<Translation>
  upstream: AppUpstream
}
// ...
type CatalogApp = {
  added_in_catalog: number
  antifeatures: string[]
  category: string
  featured: boolean
  git: {
    branch: string
    revision: string
    url: string | null
  } | null
  high_quality: boolean
  id: string
  lastUpdate: number
  level: AppLevel
  logo_hash: string | null
  maintained: boolean
  manifest: AppMinManifest
  potential_alternative_to: string[]
  state: AppState
  subtags: string[]
  repository: string
  installed: boolean
}

export type Catalog = {
  apps: Obj<CatalogApp>
  categories: {
    description: string
    icon: string
    id: string
    subtags: { id: string; title: string }[]
    title: string
  }[]
  antifeatures: {
    description: string
    icon: string
    id: string
    title: string
  }[]
}

export type AppUpgradeInfo = {
  status:
    | 'upgradable'
    | 'up_to_date'
    | 'url_required'
    | 'bad_quality'
    | 'fail_requirements'
  message: string
  url: string | null
  current_version: string
  new_version: string | null
  new_revision: string | null
  requirements: Obj | null
  specific_channel: string | null
  specific_channel_message: string | null
  notifications: Obj<string>
}

export type AppUpgradeResult = {
  notifications: Obj<Obj<Translation>>
}

export type AppInfo = {
  id: string
  description: string
  label: string
  name: string
  version: string
  domain_path: string
  logo: string | null
  screenshot?: string
  upgradable: string
  upgrade: AppUpgradeInfo
  settings: { domain?: string; path?: string } & Obj
  permissions: Obj<AppPermInfos & { sublabel: string }>
  manifest: AppMinManifest & {
    install: Obj<AnyOption>
    upstream: {
      license: string | null
      license_url: string | null
      website: string | null
      demo: string | null
      admindoc: string | null
      userdoc: string | null
      code: string | null
    }
    resources: Obj
    doc: {
      DESCRIPTION?: Translation
      ADMIN?: Translation
    } & Obj<Translation>
    notifications: {
      PRE_INSTALL: Obj<Translation> | null
      POST_INSTALL: Obj<Translation> | null
      PRE_UPGRADE: Obj<Translation> | null
      POST_UPGRADE: Obj<Translation> | null
    }
  }
  from_catalog: CatalogApp
  is_webapp: boolean
  is_default: boolean
  supports_change_url: boolean
  supports_backup_restore: boolean
  supports_multi_instance: boolean
  supports_config_panel: boolean
  supports_purge: boolean
}

export type AppList = { apps: AppInfo[] }

// BACKUP

export type BackupHookDataKeys =
  | 'data_xmpp'
  | 'data_multimedia'
  | 'data_mail'
  | 'data_home'
export type BackupHookKeys =
  | BackupHookDataKeys
  | 'conf_ynh_settings'
  | 'conf_ldap'
  | 'conf_manually_modified_files'
  | 'conf_ynh_certs'

export type BackupHooksList = {
  hooks: BackupHookKeys[]
}

export type BackupAppList = {
  apps: Pick<
    AppInfo,
    'description' | 'name' | 'version' | 'domain_path' | 'id'
  >[]
}

export type BackupInfo = {
  path: string
  created_at: string
  description: string
  size: number
  // TODO as array like everywhere else?
  apps: Obj<
    Pick<AppInfo, 'description' | 'name' | 'version'> & { size: number }
  >
  system: Record<BackupHookKeys, { paths: string[]; size: number }>
  from_yunohost_version: string
}